requirements.txt
python-gitlab
pip>=9
setuptools>=26
twine
wheel>=0.29
numpy==1.18.5

numpy
tensorflow-cpu

black
coverage
flake8
pytest
yapf
tkinter

gcsfs
google-cloud-storage
uvicorn
fastapi
requests

six>=1.14
streamlit
python-multipart